Fixes with refreturn
When wrapping a GObject* into RefPtr<T>, this preserves the reference
counter of the GObject* but the RefPtr<T> (actually a std::shared_ptr)
has a reference counter of 1. The deleter of the RefPtr<T> does
g_object_unref. This is the right behaviour when transferring the
ownership.

For methods that borrow you a reference (i.e. the object is owned
by the instance) we need to make sure the GObject*
(not the RefPtr) is not destroyed by the deleter of RefPtr. This
requires increasing the GObject* ref count first. We use
refreturn for that.