Add test-preferences-window.cc
Also fix the handling of toasts. They are owned by the methods,
so they are not used practically with Glib::RefPtr (we see
ourselves forced to use Glib::RefPtr::reference() all the time which
is awkward and easy to forget).

Let's be more honest and use plain pointers, after all this does not
prevent using Glib::RefPtr with them as long as the user remembers
to call Glib::RefPtr::reference().