Provide an implementation to wrap Gtk.SelectionModel
This way it is possible to access to the underlying Gio.ListModel.

This is similar to what gtkmm does with the SelectionModel returned
by Gtk::Stack. I would have been able to reuse Gtk::SelectionListModelImpl
class if it had not been constrained for Gtk::Stack only. So I copied
the implementation inside the Adw namespace.